% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpgaussian.default.R
\name{ahpgaussian.default}
\alias{ahpgaussian.default}
\title{Default method for ahpgaussian}
\usage{
\method{ahpgaussian}{default}(x)
}
\arguments{
\item{x}{A data frame with criteria, alternatives and a column \code{min_max}.}
}
\value{
An object of class \code{ahpgaussian}.
}
\description{
Apply the AHP-Gaussian method to a data frame.
}
