% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpgaussian.R
\name{ahpgaussian}
\alias{ahpgaussian}
\title{Analytic Hierarchy Process with Gaussian adaptation}
\usage{
ahpgaussian(x)
}
\arguments{
\item{x}{Input object (e.g., a data.frame).}
}
\value{
An object of class \code{ahpgaussian}.
}
\description{
Generic S3 function to apply the AHP-Gaussian method to different input types.
Currently implemented for data frames.
}
