% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\description{
Re-export of the magrittr pipe operator \code{\%>\%}.
}
\details{
This operator is commonly used to pass the result of one expression
as the first argument of the next function call.
}
\keyword{internal}
