% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ahpgaussian.R
\name{print.summary.ahpgaussian}
\alias{print.summary.ahpgaussian}
\title{Print method for summary.ahpgaussian}
\usage{
\method{print}{summary.ahpgaussian}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.ahpgaussian}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Compact print method for summaries of AHP Gaussian results.
}
